Moved writers.py from commonlib to baselib
