+++ openquake/baselib/writers.py
@@ -228,7 +228,7 @@ def parse_header(header):
     """
     Convert a list of the form `['fieldname:fieldtype:fieldsize',...]`
     into a numpy composite dtype. The parser understands headers generated
-    by :func:`openquake.commonlib.writers.build_header`.
+    by :func:`openquake.baselib.writers.build_header`.
     Here is an example:
 
     >>> parse_header(['PGA:float32', 'PGV', 'avg:float32:2'])
