+++ openquake/commonlib/tests/writers_test.py
@@ -21,7 +21,7 @@ import unittest
 import tempfile
 from io import BytesIO
 import psutil
-from openquake.commonlib.writers import write_csv
+from openquake.baselib.writers import write_csv
 from openquake.baselib.performance import memory_rss
 from openquake.baselib.node import Node, tostring, StreamingXMLWriter
 from xml.etree import ElementTree as etree
